requirements.txt
absl-py==0.8.0
astor==0.8.0
atari-py==0.2.6
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
coloredlogs==10.0
cycler==0.10.0
Cython==0.29.13
dill==0.3.0
future==0.17.1
gast==0.3.2
glfw==1.8.3
glob2==0.7
google-pasta==0.1.7
grpcio==1.24.0
gym==0.10.9
h5py==2.10.0
humanfriendly==4.18
idna==2.8
imageio==2.5.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.1.1
matplotlib==3.0.3
mpi4py==3.0.2
mujoco-py==1.50.1.5
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.24.2
Pillow==6.1.0
portalocker==1.5.1
protobuf==3.9.2
pybullet==2.5.5
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.36.1
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
zmq==0.0.0
stable_baselines==2.5.0
